move firebaseConfig to config.ts
